some fixes, correct pow (with comments)

+++ dump/for_dump.cpp
@@ -25,7 +25,7 @@ static void graph_create_point(Node* node, FILE* file, VariableArr* all_var)
     else // Var
     {
         char* name_var = NULL;
-        for (int i = 0; i < all_var->size; i++)
+        for (size_t i = 0; i < all_var->size; i++)
         {
             if (all_var->arr[i].num == node->value.var_num) { name_var = all_var->arr[i].name; break;}
         }
